+++ src/pages/auth/Login.jsx
@@ -1,10 +1,10 @@
 import { auth } from "../../shared/firebase";
 import { signInWithEmailAndPassword } from "firebase/auth";
-import { authErrors } from "./authErrors";
+import { authErrors } from "./constant/authErrors";
 import { FirebaseError } from "@firebase/util";
 import { Link, useNavigate } from "react-router-dom";
 import { useState } from "react";
-import { handleError } from "./handleError";
+import { handleError } from "./util/handleError";
 import logo from "/images/logo.svg";
 import {
   Wrapper,
@@ -16,8 +16,8 @@ import {
   Logo,
   Error,
   Form,
-} from "./auth-component";
-import LoginInput from "./LoginInput";
+} from "./components/auth-component";
+import LoginInput from "./components/LoginInput";
 import Button from "../../shared/components/Button";
 import PageTitle from "../../shared/components/PageTitle";
