+++ src/pages/auth/components/auth-component.jsx
